Refactor: Restructure project for PyPI packaging
- Created `tensorus` directory to house the core Python package.
- Moved relevant Python modules (`api.py`, `app.py`, `tensor_ops.py`, etc.) into the `tensorus` directory.
- Added `tensorus/__init__.py` and configured it to export `ui_utils` for easier access by Streamlit pages.
- Updated `pyproject.toml` with `tool.setuptools.packages.find` to ensure correct package discovery by `setuptools`.
- Adjusted import statements in test files and within the `tensorus` package to reflect the new structure (e.g., using relative imports or `from tensorus...`).

Note: Due to limitations with filenames containing emojis, import statements in some Streamlit `pages/` files (e.g., `pages/1_📊_Dashboard.py`) could not be automatically updated. These may require manual adjustment from `from ui_utils import ...` to `from tensorus import ...` for the UI to function correctly.
