+++ pyproject.toml
@@ -31,3 +31,8 @@ Repository = "https://github.com/tensorus/tensorus"
 [build-system]
 requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["tensorus*"]
+namespaces = false
