+++ tensorus/ingestion_agent.py
@@ -29,7 +29,7 @@ import torchvision.transforms as T # Use torchvision for image transforms
 import collections # Added import
 
 from typing import Dict, Callable, Optional, Tuple, List, Any
-from tensor_storage import TensorStorage # Import our storage module
+from .tensor_storage import TensorStorage # Import our storage module
 
 # Configure basic logging (can be customized further)
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
